refactor: Refactor package structure for Piece Editor components
