+++ app/src/main/java/net/ottercloud/sliderschrank/ui/pieceeditor/DropdownSelectorComponent.kt
@@ -26,7 +26,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.ottercloud.sliderschrank.ui
+package net.ottercloud.sliderschrank.ui.pieceeditor
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
